project_root/requirements.txt
torch
torchvision
matplotlib
pandas
pillow
streamlit
easyocr
numpy
git+https://github.com/openai/CLIP.git
